socket_app/requirements.txt
python-socketio
eventlet
websocket-client
uvicorn